feat: create User entity and update BaseEntity with Lombok
